Made it working with android
