+++ lib/screens/welcome_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import './fill_info_screen.dart';
 import '../common/misc/locale_text.dart';
+import '../common/widgets/language_switcher.dart';
 import '../common/widgets/start_button.dart';
 
 class WelcomeScreen extends StatelessWidget {
@@ -47,10 +48,7 @@ class WelcomeScreen extends StatelessWidget {
               ),
             ),
             const SizedBox(height: 15),
-            const _LanguageSwitcher(
-              width: 125,
-              height: 50,
-            ),
+            const LanguageSwitcher(width: 125, height: 35),
             const SizedBox(height: 15),
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 20.0),
@@ -64,99 +62,5 @@ class WelcomeScreen extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class _LanguageSwitcher extends StatefulWidget {
-  const _LanguageSwitcher({
-    Key? key,
-    this.width = 125,
-    this.height = 50,
-  }) : super(key: key);
-
-  final double width;
-  final double height;
-
-  @override
-  State<_LanguageSwitcher> createState() => _LanguageSwitcherState();
-}
-
-class _LanguageSwitcherState extends State<_LanguageSwitcher>
-    with TickerProviderStateMixin {
-  bool _currentLanguage = true;
-
-  late final AnimationController _slideController = AnimationController(
-    duration: const Duration(milliseconds: 500),
-    vsync: this,
-  );
-
-  @override
-  void dispose() {
-    _slideController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: _clickedChangeLanguage,
-      child: Stack(
-        alignment: AlignmentDirectional.center,
-        children: [
-          Container(
-            decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(25),
-                color: Theme.of(context).colorScheme.secondary),
-            width: widget.width,
-            height: widget.height,
-          ),
-          PositionedTransition(
-            rect: RelativeRectTween(
-              begin: RelativeRect.fromLTRB(widget.width * 0.45, 0, 0, 0),
-              end: RelativeRect.fromLTRB(0, 0, widget.width * 0.45, 0),
-            ).animate(CurvedAnimation(
-              parent: _slideController,
-              curve: Curves.easeInOut,
-            )),
-            child: Container(
-              decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(25),
-                  color: Theme.of(context).colorScheme.primary),
-            ),
-          ),
-          Positioned(
-            left: widget.width * 0.1,
-            height: widget.height,
-            child: Text(
-              'en',
-              style: Theme.of(context)
-                  .textTheme
-                  .headline4!
-                  .copyWith(fontWeight: FontWeight.bold),
-            ),
-          ),
-          Positioned(
-            right: widget.width * 0.1,
-            height: widget.height,
-            child: Text(
-              'fr',
-              style: Theme.of(context)
-                  .textTheme
-                  .headline4!
-                  .copyWith(fontWeight: FontWeight.bold),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _clickedChangeLanguage() {
-    _currentLanguage = !_currentLanguage;
-    LocaleText.of(context, listen: false)
-        .setLanguage(_currentLanguage ? 'fr' : 'en');
-
-    _currentLanguage ? _slideController.reverse() : _slideController.forward();
-    setState(() {});
   }
 }
